+++ R/files.R
@@ -224,6 +224,24 @@ read_events_sleepedfx <- function(dir, update = TRUE){
   return(stats::na.omit(events_final))
 }
 
+#' Read a stages exports from Compumedics software
+#'
+#' @param txt txt file path
+#' @param startTime Character string or date object of the hypnogram start.
+#' @return A dataframe of stages.
+#' @export
+read_stages_compumedics <- function(txt, startTime){
+  hypno <- read.table(txt, stringsAsFactors = FALSE, col.names = "event")
+  hypno$begin <- as.POSIXlt(startTime) + ((c(1:nrow(hypno))-1)*4)
+  hypno$end <- hypno$begin+4
+  hypno$event <- as.character(hypno$event)
+  hypno$event <- ifelse(hypno$event == "0","AWA",hypno$event)
+  hypno$event <- ifelse(hypno$event == "1","NREM",hypno$event)
+  hypno$event <- ifelse(hypno$event == "2","REM",hypno$event)
+  hypno$event <- ifelse(hypno$event == "?","AWA",hypno$event)
+  hypno
+}
+
 #' Read a Morpheo Data Format (MDF) directory to a list.
 #'
 #' @references P. Bouchequet, D. Jin, G. Solelhac, M. Chennaoui, D. Leger, "Morpheo Data Format (MDF), un nouveau format de données simple, robuste et performant pour stocker et analyser les enregistrements de sommeil", Médecine du Sommeil, vol. 15, n 1, p. 48/49, march 2018.